test(matching): list↔detail parity test runs in CI (Workstream A5)
Move the A5 parity assertion from tests/regression/ (ignored by jest.config.mjs)
into __tests__/economics/list-detail-tce-parity.test.ts so it runs in every CI
pass. Delete the dead duplicate.

Key improvement over the original A5 test: DA is now summed INDEPENDENTLY via
sumMatchPortDaUsd (the same mechanism the detail API uses) rather than borrowing
stored.tce_breakdown.da_usd — so a DA bug in the helper would be caught rather
than masked by circular reuse.

Assert Math.abs(storedTce - detailTce) <= 1 ($-1764 vs $-1764, delta=0 on
Hamburg→Singapore, 55000 DWT, 50000 MT BULK fixture).
